A simple implementation of a dependencies override mechanism for browserify transforms.
This change enables you to add a block to your top-level `package.json`:
```
"browserify": {
  "dependencies": {
    "js-csp": "transform": ["babelify"]
  }
}
```
that will apply to submodules.  The intended implementation will completely
replace any transforms in the submodule.  It also works on a 'first wins'
basis, such that on a top-down walk of the dependencies the first override
is the only one that gets used.  Also if the module is reached before any
overrides are set, none can subsequently be applied, for a consistent result!